разделил код на логические модули
